+++ Documentation/Release Notes.md
@@ -34,7 +34,7 @@ Changes:
 
 ## 0.1
 
-* Added <tile_attribution> parameter for background tiles of geoplots
+* Added <tile_attribution> & <tile_alpha> parameter for background tiles of geoplots
 * WebGL now as plotting backend as default (plots and geoplots)
 * Code Refactoring
 * Toolbar now visible per default